+++ compilers/qvis/qvis3.cpp
@@ -61,13 +61,13 @@ NewWinding
 */
 winding_t* NewWinding(int points)
 {
-	winding_t* w;
-	int			size;
+	winding_t	*w;
+	intptr_t	size;
 
 	if (points > MAX_POINTS_ON_WINDING)
 		Error("NewWinding: %i points", points);
 
-	size = (int)((winding_t*)0)->points[points];
+	size = (intptr_t)((winding_t*)0)->points[points];
 	w = (winding_t*)malloc(size);
 	memset(w, 0, size);
